minor: refactor decodeBatches to make private in broadcast exchange (#1195)
